Move the escape logic to avoid TypeError
Escape logic = the logic responsible for handling what happens when
the user presses esc in the prompt aka passes null to the prompt.

Keeping the escape logic in the playSingleRound function meant that
null could be passed to a string method, resulting in a TypeError.

Moving the escape logic to the game function fixes the error,
improves readability, and simplifies the code.